Audio/DevIchAc97.cpp: Warnings.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@67657 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Audio/DevIchAc97.cpp
@@ -1006,7 +1006,7 @@ static DECLCALLBACK(int) ichac97StreamAsyncIOThread(RTTHREAD hThreadSelf, void *
                 case AC97SOUNDSOURCE_PI_INDEX:
                 case AC97SOUNDSOURCE_MC_INDEX:
                 {
-                    cbToProcess = RTCircBufFree(pCircBuf);
+                    cbToProcess = (uint32_t)RTCircBufFree(pCircBuf);
                     if (cbToProcess)
                         rc2 = ichac97StreamWrite(pThis, pStream, pMixSink, (uint32_t)cbToProcess, &cbProcessed);
                     break;
@@ -1015,7 +1015,7 @@ static DECLCALLBACK(int) ichac97StreamAsyncIOThread(RTTHREAD hThreadSelf, void *
                 /* Output. */
                 case AC97SOUNDSOURCE_PO_INDEX:
                 {
-                    cbToProcess = RTCircBufUsed(pCircBuf);
+                    cbToProcess = (uint32_t)RTCircBufUsed(pCircBuf);
                     if (cbToProcess)
                         rc2 = ichac97StreamRead(pThis, pStream, pMixSink, (uint32_t)cbToProcess, &cbProcessed);
                     break;
